refactor schemas
