+++ backend/app/api/api_v1/routers/auth.py
@@ -10,9 +10,7 @@ auth_router = r = APIRouter()
 
 
 @r.post("/token")
-async def login(
-    db=Depends(get_db), form_data: OAuth2PasswordRequestForm = Depends()
-):
+async def login(db=Depends(get_db), form_data: OAuth2PasswordRequestForm = Depends()):
     user = authenticate_user(db, form_data.username, form_data.password)
     if not user:
         raise HTTPException(
@@ -21,9 +19,7 @@ async def login(
             headers={"WWW-Authenticate": "Bearer"},
         )
 
-    access_token_expires = timedelta(
-        minutes=security.ACCESS_TOKEN_EXPIRE_MINUTES
-    )
+    access_token_expires = timedelta(minutes=security.ACCESS_TOKEN_EXPIRE_MINUTES)
     if user.is_superuser:
         permissions = "admin"
     else:
@@ -37,9 +33,7 @@ async def login(
 
 
 @r.post("/signup")
-async def signup(
-    db=Depends(get_db), form_data: OAuth2PasswordRequestForm = Depends()
-):
+async def signup(db=Depends(get_db), form_data: OAuth2PasswordRequestForm = Depends()):
     user = sign_up_new_user(db, form_data.username, form_data.password)
     if not user:
         raise HTTPException(
@@ -48,9 +42,7 @@ async def signup(
             headers={"WWW-Authenticate": "Bearer"},
         )
 
-    access_token_expires = timedelta(
-        minutes=security.ACCESS_TOKEN_EXPIRE_MINUTES
-    )
+    access_token_expires = timedelta(minutes=security.ACCESS_TOKEN_EXPIRE_MINUTES)
     if user.is_superuser:
         permissions = "admin"
     else:
