+++ backend/app/api/api_v1/routers/tests/test_auth.py
@@ -43,15 +43,11 @@ def test_resignup(client, test_user, monkeypatch):
     assert response.status_code == 409
 
 
-def test_wrong_password(
-    client, test_db, test_user, test_password, monkeypatch
-):
+def test_wrong_password(client, test_db, test_user, test_password, monkeypatch):
     def verify_password_failed_mock(first: str, second: str):
         return False
 
-    monkeypatch.setattr(
-        security, "verify_password", verify_password_failed_mock
-    )
+    monkeypatch.setattr(security, "verify_password", verify_password_failed_mock)
 
     response = client.post(
         "/api/token", data={"username": test_user.email, "password": "wrong"}
