+++ backend/app/core/auth.py
@@ -2,13 +2,15 @@ import jwt
 from fastapi import Depends, HTTPException, status
 from jwt import PyJWTError
 
+import app.db.models.user
+import app.db.schemas.user
 from app.core import security
-from app.db import models, schemas, session
-from app.db.crud import get_user_by_email, create_user
+from app.db import session
+from app.db.crud.user import get_user_by_email, create_user
 
 
 async def get_current_user(
-        db=Depends(session.get_db), token: str = Depends(security.oauth2_scheme)
+    db=Depends(session.get_db), token: str = Depends(security.oauth2_scheme)
 ):
     credentials_exception = HTTPException(
         status_code=status.HTTP_401_UNAUTHORIZED,
@@ -23,7 +25,7 @@ async def get_current_user(
         if email is None:
             raise credentials_exception
         permissions: str = payload.get("permissions")
-        token_data = schemas.TokenData(email=email, permissions=permissions)
+        token_data = app.db.schemas.user.TokenData(email=email, permissions=permissions)
     except PyJWTError:
         raise credentials_exception
     user = get_user_by_email(db, token_data.email)
@@ -33,7 +35,7 @@ async def get_current_user(
 
 
 async def get_current_active_user(
-        current_user: models.User = Depends(get_current_user),
+    current_user: app.db.models.user.User = Depends(get_current_user),
 ):
     if not current_user.is_active:
         raise HTTPException(status_code=400, detail="Inactive user")
@@ -41,8 +43,8 @@ async def get_current_active_user(
 
 
 async def get_current_active_superuser(
-        current_user: models.User = Depends(get_current_user),
-) -> models.User:
+    current_user: app.db.models.user.User = Depends(get_current_user),
+) -> app.db.models.user.User:
     if not current_user.is_superuser:
         raise HTTPException(
             status_code=403, detail="The user doesn't have enough privileges"
@@ -65,7 +67,7 @@ def sign_up_new_user(db, email: str, password: str):
         return False  # User already exists
     new_user = create_user(
         db,
-        schemas.UserCreate(
+        app.db.schemas.user.UserCreate(
             email=email,
             password=password,
             is_active=True,
