+++ backend/app/initial_data.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
 
-from app.db.crud import create_user
-from app.db.schemas import UserCreate
-from app.db.session import SessionLocal
-
 import os
+
+from app.db.crud.user import create_user
+from app.db.schemas.user import UserCreate
+from app.db.session import SessionLocal
 
 
 def create_superuser(email, password):
@@ -23,7 +23,10 @@ def create_superuser(email, password):
 
 def init() -> None:
     create_superuser(os.getenv("SUPERUSER_EMAIL"), os.getenv("SUPERUSER_PASSWORD"))
-    create_superuser(os.getenv("MACHINE_USER_LOADER_EMAIL"), os.getenv("MACHINE_USER_LOADER_PASSWORD"))
+    create_superuser(
+        os.getenv("MACHINE_USER_LOADER_EMAIL"),
+        os.getenv("MACHINE_USER_LOADER_PASSWORD"),
+    )
 
 
 if __name__ == "__main__":
