+++ loader/app/main.py
@@ -76,8 +76,8 @@ if __name__ == "__main__":
 
     # wait for backend
     # TODO more robust solution here: https://github.com/climatepolicyradar/navigator/issues/368
-    logger.warning("Waiting 10 seconds for backend...")
-    time.sleep(10)
+    logger.warning("Waiting 5 seconds for backend...")
+    time.sleep(0)
     logger.warning("Wait complete. Starting...")
 
     main()
